requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
PyYAML>=6.0
tqdm>=4.62.0

# API and token management dependencies
openai>=1.0.0
tiktoken>=0.5.0

# Optional dependencies for plotting
matplotlib>=3.5.0
seaborn>=0.11.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
